Files
I think I got everything writing to and reading from the file as necessary. All the data is stored... until the app is paused. If it is stopped I think I keep the data, but if it's Paused or Destroyed everything is lost. I just need to Override those functions.

Also, make sure the contentView actually displays all the contents of ArrayList<Day> days, it seems that the contents don't show up after booting the app back up, even if the data was actually saved.