Fix Issues with Files
Data now saves to file. It appears to be written properly to the file, and read properly from the file. All Day objects are displayed correctly on the MainActivity.

The next step would be to add a way to export data to an Excel file format, and clear the storage file. Possibly make sure that only two weeks at a time are exported, and cleared from the file. Everything else should stay and be rewritten to the app file.