Add Ability to Export Punches to Excel Format
BEFORE: Find a suitable format that can be read by Excel, write function to make writing it possible, add a button in the menu to Export. Also need to make sure the exported Punches are removed from the "days" ArrayList and removed from the app data file.

After: I had almost everything working through exporting the document as an email attachment... but then it turns out that Gmail has some kind of security measures that I can't get around. Anyway, the file MAY be exportable, but not through Gmail.