More CSV
BEFORE: I need to find a way to make a .csv file that can actually be exported in some way. Security seems to be the main problem right now, so I might need to export as the body of an email and make the .csv file on the computer. A very ghetto solution, but it would get the job done. And then I could move on to Monkey See!

AFTER: ...so pretty much nothing changed. So there's that.